Replace interfaces with types
this will probably be partially reverted in the future, but right now anything that gets passed into `RESTManager#api()` needs to be a type alias because interfaces don't have implicit index signatures. This will probably be partially reverted in the future if I ever get a solid grasp of what interfaces are for and when/how to use them.